Add ResultScreen.cs parts
