+++ Screens/Result/ResultGraph.cs
@@ -1,0 +1,44 @@
+// Copyright (c) mk56_spn <dhsjplt@gmail.com>.Licensed under the GNU General Public Licence (2.0).
+// See the LICENCE file in the repository root for full licence text.
+
+using System;
+using System.Linq;
+using Friflo.Engine.ECS;
+using Godot;
+using XanaduProject.DataStructure;
+using XanaduProject.ECSComponents;
+using XanaduProject.ECSComponents.Tag;
+using XanaduProject.Tools;
+
+namespace XanaduProject.Screens.Result
+{
+    public partial class ResultGraph(EntityStore store) : PanelContainer
+    {
+
+        public override void _Draw()
+        {
+            CustomMinimumSize = new Vector2(900, 300);
+
+            DrawSetTransform(new Vector2(0,CustomMinimumSize.Y / 2));
+
+            foreach (var judgement in Enum.GetValues<Judgement>().Reverse())
+            {
+                var color = JudgementInfo.GetJudgmentColor(judgement).Darkened(0.3f);
+                float offset = (float)JudgementInfo.JudgementDeviation(judgement);
+                DrawRect(new Rect2(new Vector2(0, -offset), new Vector2(CustomMinimumSize.X - 30, offset * 2)), color.Darkened(0.8f));
+                DrawLine(new Vector2(0, offset), new Vector2(CustomMinimumSize.X,offset), color);
+                DrawLine(new Vector2(0, -offset), new Vector2(CustomMinimumSize.X,-offset), color);
+            }
+            DrawLine(Vector2.Zero, new Vector2(CustomMinimumSize.X, 0), Colors.White);
+            store.Query<Judged, NoteEcs>()
+                .ForEachEntity((ref Judged judged,ref NoteEcs note, Entity _) =>
+                {
+                    DrawRect(new Rect2(new Vector2(note.TimingPoint *4,judged.Deviation), new Vector2(5,5)), JudgementInfo.GetJudgmentColor(judged.Judgement));
+                });
+
+            DrawSetTransform(default);
+            DrawRect(new Rect2(Vector2.Zero, CustomMinimumSize), XanaduColors.XanaduYellow, false, 3);
+
+        }
+    }
+}
